+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Changelog
 
 **Next release**
 
+* Added IPC restriction *ICameraService*
+
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)
 
 **Version 2.1.24-1 TEST**